Tokenizers.DecodeStream (Tokenizers v0.6.0-dev)

Implements streaming decoding functionality for tokenizers.

Types

t()

@type t() :: %Tokenizers.DecodeStream{resource: reference()}

Functions

info(decode_stream)

Returns information about the decode stream state.

new(opts \\ [])

@spec new(keyword()) :: t()

Creates a new decode stream.

Options

:skip_special_tokens - determines whether special tokens should be skipped during decoding. By default, it is set to false.

step(decode_stream, tokenizer, id)

Steps through the decode stream with the given tokenizer and token ID.

Returns {:ok, String.t()} if there's a decoded string, or {:ok, :out_ofr_range} if the token ID is out of range. Returns {:error, reason} if an error occurs during decoding.