Tokenizers.Decoder (Tokenizers v0.3.2)

The Decoder knows how to go from the IDs used by the Tokenizer, back to a readable piece of text. Some Normalizer and PreTokenizer use special characters or identifiers that need to be reverted.

Types

bpe_options()

@type bpe_options() :: [{:suffix, String.t()}]

Options for BPE decoder initialization. All options can be ommited.

suffix - The suffix to add to the end of each word, defaults to </w>

ctc_options()

@type ctc_options() :: [
  pad_token: String.t(),
  word_delimiter_token: String.t(),
  cleanup: boolean()
]

Options for CTC decoder initialization. All options can be ommited.

pad_token - The token used for padding, defaults to <pad>

word_delimiter_token - The token used for word delimiter, defaults to |

cleanup - Whether to cleanup tokenization artifacts, defaults to true

metaspace_options()

@type metaspace_options() :: [replacement: char(), add_prefix_space: boolean()]

Options for Metaspace decoder initialization. All options can be ommited.

replacement - The replacement character, defaults to ▁ (as char)

add_prefix_space - Whether to add a space to the first word, defaults to true

t()

@type t() :: %Tokenizers.Decoder{resource: reference()}

word_piece_options()

@type word_piece_options() :: [prefix: String.t(), cleanup: boolean()]

Options for WordPiece decoder initialization. All options can be ommited.

prefix - The prefix to use for subwords, defaults to ##

cleanup - Whether to cleanup tokenization artifacts, defaults to true

Functions

bpe(options \\ [])

@spec bpe(bpe_options :: bpe_options()) :: t()

Creates new BPE decoder

byte_fallback()

@spec byte_fallback() :: t()

Creates new ByteFallback decoder

byte_level()

@spec byte_level() :: t()

Creates new ByteLevel decoder

ctc(options \\ [])

@spec ctc(ctc_options :: ctc_options()) :: t()

Creates new CTC decoder

decode(decoder, tokens)

@spec decode(t(), [String.t()]) :: {:ok, String.t()} | {:error, any()}

Decodes tokens into string with provided decoder.

fuse()

@spec fuse() :: t()

Creates new Fuse decoder

metaspace(options \\ [])

@spec metaspace(metaspace_options :: metaspace_options()) :: t()

Creates new Metaspace decoder

replace(pattern, content)

@spec replace(pattern :: String.t(), content :: String.t()) :: t()

Creates new Replace decoder

sequence(decoders)

@spec sequence(decoders :: [t()]) :: t()

Creates new Sequence decoder

strip(content, left, right)

@spec strip(content :: char(), left :: non_neg_integer(), right :: non_neg_integer()) ::
  t()

Creates new Strip decoder.

It expects a character and the number of times to strip the character on left and right sides.

word_piece(options \\ [])

@spec word_piece(word_piece_options :: word_piece_options()) :: t()

Creates new WordPiece decoder