Tokenizers.Encoding.Transformation (Tokenizers v0.6.0-dev)

Module containing handy functions to build the transformations list.

This list is applied to an encoding using Tokenizers.Encoding.transform/2.

Types

t()

@type t() :: [
  pad: {non_neg_integer(), Tokenizers.Encoding.padding_opts()},
  truncate: {non_neg_integer(), Tokenizers.Encoding.truncation_opts()},
  set_sequence_id: non_neg_integer()
]

Functions

pad(target_length, opts \\ [])

@spec pad(non_neg_integer(), Tokenizers.Encoding.padding_opts()) ::
  {:pad, {non_neg_integer(), Tokenizers.Encoding.padding_opts()}}

Generates the padding transformation.

Check Tokenizers.Encoding.pad/3 for more information.

set_sequence_id(id)

@spec set_sequence_id(non_neg_integer()) :: {:set_sequence_id, non_neg_integer()}

Generates the set_sequence_id transformation.

Check Tokenizers.Encoding.set_sequence_id/2 for more information.

truncate(max_length, opts \\ [])

@spec truncate(non_neg_integer(), Tokenizers.Encoding.truncation_opts()) ::
  {:truncate, {non_neg_integer(), Tokenizers.Encoding.truncation_opts()}}

Generates the truncation transformation.

Check Tokenizers.Encoding.truncate/3 for more information.