Tokenizers.Model (Tokenizers v0.3.2)

The struct and associated functions for the tokenizer model.

Types

save_opts()

@type save_opts() :: [{:prefix, String.t()}]

Options to save the model. All options can be ommited.

:prefix (default "") - The prefix to use for all the files that will get created.

t()

@type t() :: %Tokenizers.Model{resource: reference()}

Represents different kind of models that can be used across the library.

Functions

info(model)

@spec info(model :: t()) :: map()

Retrieves information about the model.

Information retrieved differs per model but all include model_type.

save(model, folder, opts \\ [])

@spec save(model :: t(), folder :: String.t(), opts :: save_opts()) ::
  {:ok, file_paths :: [String.t()]} | {:error, any()}

Save the current model in the given folder, using the given name for the various files that will get created. Any file with the same name that already exist in this folder will be overwritten.