Tokenizers.Model.BPE (Tokenizers v0.3.2)

Types

options()

@type options() :: [
  cache_capacity: number(),
  dropout: float(),
  unk_token: String.t(),
  continuing_subword_prefix: String.t(),
  end_of_word_suffix: String.t(),
  fuse_unk: boolean(),
  byte_fallback: boolean()
]

Options for model initialisation. All options can be ommited.

:cache_capacity (default 10_000) - The number of words that the BPE cache can contain. The cache allows to speed-up the process by keeping the result of the merge operations for a number of words.

:dropout - The BPE dropout to use. Must be an float between 0 and 1

:unk_token - The unknown token to be used by the model

:continuing_subword_prefix - The prefix to attach to subword units that don't represent a beginning of word

:end_of_word_suffix - The suffix to attach to subword units that represent an end of word

Functions

empty()

@spec empty() :: {:ok, Tokenizers.Model.t()}

Instantiate an empty BPE Model

from_file(vocab, merges, options \\ [])

@spec from_file(
  vocab :: String.t(),
  merges :: String.t(),
  options :: options()
) :: {:ok, Tokenizers.Model.t()}

Instantiate a BPE model from the given vocab and merges files

init(vocab, merges, options \\ [])

@spec init(
  vocab :: %{required(String.t()) => integer()},
  merges :: [{String.t(), String.t()}],
  options :: options()
) :: {:ok, Tokenizers.Model.t()}

Instantiate a BPE model from the given vocab and merges