Tokenizers.Normalizer (Tokenizers v0.4.0-dev)

A Normalizer is in charge of pre-processing the input string in order to normalize it as relevant for a given use case.

Some common examples of normalization are the Unicode normalization algorithms (NFD, NFKD, NFC & NFKC), lowercasing etc... The specificity of tokenizers is that we keep track of the alignment while normalizing. This is essential to allow mapping from the generated tokens back to the input text.

The Normalizer is optional.

bert_opts()

@type bert_opts() :: [
  clean_text: boolean(),
  handle_chinese_chars: boolean(),
  strip_accents: boolean(),
  lowercase: boolean()
]

Options for BERT normalizer initialisation. All values are optional.

:clean_text (default true) - Whether to clean the text, by removing any control characters and replacing all whitespaces by the classic one.

:handle_chinese_chars (default true) - Whether to handle chinese chars by putting spaces around them.

:strip_accents - Whether to strip all accents. If this option is not specified, then it will be determined by the value for lowercase (as in the original Bert).

:lowercase (default true) - Whether to lowercase.

strip_opts()

@type strip_opts() :: [left: boolean(), right: boolean()]

Options for Strip normalizer initialisation. All values are optional.

:left (default true) - Whether to strip left side.

:right (default true) - Whether to strip right side.

t()

@type t() :: %Tokenizers.Normalizer{resource: reference()}

bert_normalizer(opts \\ [])

@spec bert_normalizer(opts :: bert_opts()) :: t()

Takes care of normalizing raw text before giving it to a Bert model. This includes cleaning the text, handling accents, chinese chars and lowercasing.

lowercase()

@spec lowercase() :: t()

Replaces all uppercase to lowercase

nfc()

@spec nfc() :: t()

NFC Unicode Normalizer

nfd()

@spec nfd() :: t()

NFD Unicode Normalizer,

nfkc()

@spec nfkc() :: t()

NFKC Unicode Normalizer

nfkd()

@spec nfkd() :: t()

NFKD Unicode Normalizer

nmt()

@spec nmt() :: t()

Nmt normalizer

normalize(normalizer, input)

@spec normalize(normalizer :: t(), input :: String.t()) :: {:ok, String.t()}

Normalizes the input presented as string into new string

precompiled(data)

@spec precompiled(data :: binary()) :: {:ok, t()} | {:error, any()}

Precompiled normalizer. Don’t use manually it is used for compatiblity for SentencePiece.

prepend(prepend)

@spec prepend(prepend :: String.t()) :: t()

Prepend normalizer.

replace(pattern, content)

@spec replace(pattern :: String.t(), content :: String.t()) :: t()

Replaces a custom string or regexp and changes it with given content

sequence(normalizers)

@spec sequence(normalizers :: [t()]) :: t()

Composes multiple normalizers that will run in the provided order.

strip(opts \\ [])

@spec strip(opts :: strip_opts()) :: t()

Strip normalizer. Removes all whitespace characters on the specified sides (left, right or both) of the input

strip_accents()

@spec strip_accents() :: t()

Strip Accent normalizer. Removes all accent symbols in unicode (to be used with NFD for consistency).