Tokenizers.PostProcessor (Tokenizers v0.4.0-dev)

After the whole pipeline, we sometimes want to insert some special tokens before feed a tokenized string into a model like ”[CLS] My horse is amazing [SEP]”. The PostProcessor is the component doing just that.

byte_level_opts()

@type byte_level_opts() :: [{:trim_offsets, boolean()}]

Options for ByteLevel post-processor. All values are optional.

:trim_offsets (default true) - Whether to trim the whitespaces in the produced offsets

roberta_opts()

@type roberta_opts() :: [trim_offsets: boolean(), add_prefix_space: boolean()]

Options for Roberta post-processor. All values are optional.

:trim_offest (default true) - Whether to trim the whitespaces in the produced offsets

:add_prefix_space (default true) - Whether add_prefix_space was ON during the pre-tokenization.

t()

@type t() :: %Tokenizers.PostProcessor{resource: reference()}

template_opts()

@type template_opts() :: [
  single: String.t(),
  pair: String.t(),
  special_tokens: [{String.t(), integer()}]
]

Options for Template post-processor.

:single - A string describing the template for a single sequence.

:pair - A string describing the template for a pair of sequences.

:special_tokens - A list of special tokens to use in the template.

bert(sep, cls)

@spec bert(
  sep :: {String.t(), integer()},
  cls :: {String.t(), integer()}
) :: t()

Instantiate a new BertProcessing with the given tokens

Params are tuple with the string representation of the token, and its id

byte_level(opts \\ [])

@spec byte_level(opts :: byte_level_opts()) :: t()

Creates ByteLevel post-processor.

roberta(sep, cls, opts \\ [])

@spec roberta(
  sep :: {String.t(), integer()},
  cls :: {String.t(), integer()},
  opts :: roberta_opts()
) :: t()

Creates Roberta post-processor.

sequence(post_processors)

@spec sequence(post_processors :: [t()]) :: t()

Instantiate a new Sequence post-processor

template(opts \\ [])

@spec template(opts :: template_opts()) :: t()

Creates Template post-processor.

Let’s you easily template the post processing, adding special tokens, and specifying the type_id for each sequence/special token. The template is given two strings representing the single sequence and the pair of sequences, as well as a set of special tokens to use.

Example, when specifying a template with these values:

single: "[CLS] $A [SEP]"

pair: "[CLS] $A [SEP] $B [SEP]"

special tokens:

"[CLS]"

"[SEP]"

Input: ("I like this", "but not this") Output: "[CLS] I like this [SEP] but not this [SEP]"