Tokenizers.Trainer (Tokenizers v0.4.0-dev)

A Trainer has the responsibility to train a model. We feed it with lines/sentences and then it can train the given Model.

bpe_options()

@type bpe_options() :: [
  vocab_size: non_neg_integer(),
  min_frequency: non_neg_integer(),
  special_tokens: [String.t()],
  limit_alphabet: non_neg_integer(),
  initial_alphabet: [char()],
  show_progress: boolean(),
  continuing_subword_prefix: String.t(),
  end_of_word_suffix: String.t()
]

Options for BPE trainer initialisation. All options can be ommited.

t()

@type t() :: %Tokenizers.Trainer{resource: reference()}

unigram_options()

@type unigram_options() :: [
  vocab_size: non_neg_integer(),
  n_sub_iterations: non_neg_integer(),
  shrinking_factor: float(),
  special_tokens: [String.t()],
  initial_alphabet: [char()],
  uni_token: String.t(),
  max_piece_length: non_neg_integer(),
  seed_size: non_neg_integer(),
  show_progress: boolean()
]

Options for Unigram trainer initialisation. All options can be ommited.

wordlevel_options()

@type wordlevel_options() :: [
  vocab_size: non_neg_integer(),
  min_frequency: non_neg_integer(),
  special_tokens: [String.t()],
  show_progress: boolean()
]

Options for WordLevel trainer initialisation. All options can be ommited.

wordpiece_options()

@type wordpiece_options() :: [
  vocab_size: non_neg_integer(),
  min_frequency: non_neg_integer(),
  special_tokens: [String.t()],
  limit_alphabet: non_neg_integer(),
  initial_alphabet: [char()],
  show_progress: boolean(),
  continuing_subword_prefix: String.t(),
  end_of_word_suffix: String.t()
]

Options for WordPiece trainer initialisation. All options can be ommited.

bpe(options \\ [])

@spec bpe(bpe_options()) :: {:ok, t()} | {:error, any()}

Creates a new BPE Trainer.

info(trainer)

@spec info(t()) :: map()

Get trainer info

train(trainer, model)

@spec train(t(), Tokenizers.Model.t()) :: {:ok, [String.t()]} | {:error, any()}

The actual training method. This will mutate a Model as well as return a list of special_tokens to be added directly to the tokenizer along with the model.

unigram(options \\ [])

@spec unigram(unigram_options()) :: {:ok, t()} | {:error, any()}

Creates a new Unigram Trainer.

wordlevel(options \\ [])

@spec wordlevel(wordlevel_options()) :: {:ok, t()} | {:error, any()}

Creates a new WordLevel Trainer.

wordpiece(options \\ [])

@spec wordpiece(wordpiece_options()) :: {:ok, t()} | {:error, any()}

Creates a new WordPiece Trainer.