Tokenizers (Tokenizers v0.1.0-dev)

Elixir bindings to Hugging Face Tokenizers.

Hugging Face describes the Tokenizers library as:

Fast State-of-the-art tokenizers, optimized for both research and production

🤗 Tokenizers provides an implementation of today’s most used tokenizers, with a focus on performance and versatility. These tokenizers are also used in 🤗 Transformers.

This library has bindings to use pretrained tokenizers. Support for building and training a tokenizer from scratch is forthcoming.

Functions

decode(tokenizer, ids)

Specs

decode(Tokenizers.Tokenizer.t(), binary() | [binary()]) ::
  {:ok, Tokenizers.Encoding.t() | [Tokenizers.Encoding.t()]} | {:error, term()}

Decode the given list of ids or list of lists of ids back to strings.

encode(tokenizer, input)

Specs

encode(Tokenizers.Tokenizer.t(), binary() | [binary()]) ::
  Tokenizers.Encoding.t() | [Tokenizers.Encoding.t()]

Encode the given sequence or batch of sequences to ids.

from_file(path)

Specs

from_file(binary()) :: Tokenizers.Tokenizer.t()

Instantiate a new tokenizer from the file at the given path.

from_pretrained(identifier)

Specs

from_pretrained(binary()) :: Tokenizers.Tokenizer.t()

Instantiate a new tokenizer from an existing file on the Hugging Face Hub.

get_attention_mask(encoding)

Specs

get_attention_mask(Tokenizers.Encoding.t()) ::
  {:ok, [integer()]} | {:error, term()}

Get the attention_mask from an encoding.

get_ids(encoding)

Specs

get_ids(Tokenizers.Encoding.t()) :: {:ok, [integer()]} | {:error, term()}

Get the ids from an encoding.

get_tokens(encoding)

Specs

get_tokens(Tokenizers.Encoding.t()) :: {:ok, [binary()]} | {:error, term()}

Get the tokens from an encoding.

get_vocab(tokenizer)

Specs

get_vocab(Tokenizers.Tokenizer.t()) ::
  {:ok, %{required(binary()) => integer()}} | {:error, term()}

Get the tokenizer's vocabulary as a map of token to id.

get_vocab_size(tokenizer)

Specs

get_vocab_size(Tokenizers.Tokenizer.t()) :: {:ok, integer()} | {:error, term()}

Get the number of tokens in the vocabulary.

id_to_token(tokenizer, id)

Specs

id_to_token(Tokenizers.Tokenizer.t(), integer()) ::
  {:ok, binary()} | {:error, term()}

Convert a given id to its token.

pad(encoding, length, opts \\ [])

Specs

pad(
  encoding :: Tokenizers.Encoding.t(),
  length :: pos_integer(),
  opts :: Keyword.t()
) ::
  Tokenizers.Encoding.t()

Pad the encoding to the given length.

Options

direction - The padding direction. Can be :right or :left. Default: :right.

pad_id - The id corresponding to the padding token. Default: 0.

pad_token - The padding token to use. Default: "[PAD]".

pad_type_id - The type ID corresponding to the padding token. Default: 0.

save(tokenizer, path)

Specs

save(Tokenizers.Tokenizer.t(), binary()) :: term()

Save the tokenizer to the provided path.

token_to_id(tokenizer, token)

Specs

token_to_id(Tokenizers.Tokenizer.t(), binary()) ::
  {:ok, integer()} | {:error, term()}

Convert a given token to its id.

truncate(encoding, max_len, opts \\ [])

Specs

truncate(
  encoding :: Tokenizers.Encoding.t(),
  length :: integer(),
  opts :: Keyword.t()
) ::
  {:ok, Tokenizers.Encoding.t()} | {:error, term()}

Truncate the encoding to the given length.

Options

direction - The truncation direction. Can be :right or :left. Default: :right.

stride - The length of previous content to be included in each overflowing piece. Default: 0.