Tokenizers.Decoder (Tokenizers v0.5.0-dev)

Decoders and decoding functions.

Decoder transforms a sequence of token ids back to a readable piece of text.

Some normalizers and pre-tokenizers use special characters or identifiers that need special logic to be reverted.

t()

@type t() :: %Tokenizers.Decoder{resource: reference()}

bpe(opts \\ [])

@spec bpe(keyword()) :: t()

Creates a BPE decoder.

Options

suffix - the suffix to add to the end of each word. Defaults to </w>

byte_fallback()

@spec byte_fallback() :: t()

Creates a ByteFallback decoder.

byte_level()

@spec byte_level() :: t()

Creates a ByteLevel decoder.

ctc(opts \\ [])

@spec ctc(keyword()) :: t()

Creates a CTC decoder.

Options

pad_token - the token used for padding. Defaults to <pad>

word_delimiter_token - the token used for word delimiter. Defaults to |

cleanup - whether to cleanup tokenization artifacts, defaults to true

decode(decoder, tokens)

@spec decode(t(), [String.t()]) :: {:ok, String.t()} | {:error, any()}

Decodes tokens into string with provided decoder.

fuse()

@spec fuse() :: t()

Creates a Fuse decoder.

metaspace(opts \\ [])

@spec metaspace(keyword()) :: t()

Creates a Metaspace decoder.

Options

replacement - the replacement character. Defaults to ▁ (as char)

add_prefix_space - whether to add a space to the first word. Defaults to true

replace(pattern, content)

@spec replace(String.t(), String.t()) :: t()

Creates a Replace decoder.

sequence(decoders)

@spec sequence(decoders :: [t()]) :: t()

Combines a list of decoders into a single sequential decoder.

strip(content, left, right)

@spec strip(char(), non_neg_integer(), non_neg_integer()) :: t()

Creates a Strip decoder.

It expects a character and the number of times to strip the character on left and right sides.

word_piece(opts \\ [])

@spec word_piece(keyword()) :: t()

Creates a WordPiece decoder.

Options

prefix - The prefix to use for subwords. Defaults to ##

cleanup - Whether to cleanup tokenization artifacts. Defaults to true