Tokenizers.Encoding (Tokenizers v0.3.2)

The struct and associated functions for an encoding, the output of a tokenizer.

Use these functions to retrieve the inputs needed for a natural language processing machine learning model.

Types

padding_opts()

@type padding_opts() :: [
  pad_id: non_neg_integer(),
  pad_type_id: non_neg_integer(),
  pad_token: String.t(),
  direction: :left | :right
]

Options for padding. All options can be ommited.

direction (default :right) - The padding direction.

pad_id (default 0) - The id corresponding to the padding token.

pad_type_id (default 0) - The type ID corresponding to the padding token.

pad_token (default [PDA]) - The padding token to use.

t()

@type t() :: %Tokenizers.Encoding{resource: reference()}

truncation_opts()

@type truncation_opts() :: [stride: non_neg_integer(), direction: :left | :right]

Options for truncation. All options can be ommited.

stride (default 0) - The length of previous content to be included in each overflowing piece.

direction (default :right) - The truncation direction.

Functions

char_to_token(encoding, position, seq_id)

@spec char_to_token(Encoding.t(), non_neg_integer(), non_neg_integer()) ::
  non_neg_integer() | nil

Get the token that contains the given char.

char_to_word(encoding, position, seq_id)

@spec char_to_word(Encoding.t(), non_neg_integer(), non_neg_integer()) ::
  non_neg_integer() | nil

Get the word that contains the given char.

get_attention_mask(encoding)

@spec get_attention_mask(Encoding.t()) :: [integer()]

Get the attention mask from an encoding.

get_ids(encoding)

@spec get_ids(Encoding.t()) :: [integer()]

Get the ids from an encoding.

get_length(encoding)

@spec get_length(Encoding.t()) :: non_neg_integer()

Get the number of tokens in an encoding.

get_n_sequences(encoding)

@spec get_n_sequences(Encoding.t()) :: non_neg_integer()

Return the number of sequences combined in this Encoding

get_offsets(encoding)

@spec get_offsets(Encoding.t()) :: [{integer(), integer()}]

Get offsets from an encoding.

The offsets are expressed in terms of UTF-8 bytes.

get_overflowing(encoding)

@spec get_overflowing(Encoding.t()) :: [Encoding.t()]

Get the overflow from an encoding.

get_sequence_ids(encoding)

@spec get_sequence_ids(Encoding.t()) :: [non_neg_integer() | nil]

Get sequence ids from an encoding.

get_special_tokens_mask(encoding)

@spec get_special_tokens_mask(Encoding.t()) :: [integer()]

Get the special tokens mask from an encoding.

get_tokens(encoding)

@spec get_tokens(Encoding.t()) :: [binary()]

Get the tokens from an encoding.

get_type_ids(encoding)

@spec get_type_ids(Encoding.t()) :: [integer()]

Get token type ids from an encoding.

get_u32_attention_mask(encoding)

@spec get_u32_attention_mask(Encoding.t()) :: binary()

Same as get_attention_mask/1, but returns binary with u32 values.

get_u32_ids(encoding)

@spec get_u32_ids(Encoding.t()) :: binary()

Same as get_ids/1, but returns binary with u32 values.

get_u32_special_tokens_mask(encoding)

@spec get_u32_special_tokens_mask(Encoding.t()) :: binary()

Same as get_special_tokens_mask/1, but returns binary with u32 values.

get_u32_type_ids(encoding)

@spec get_u32_type_ids(Encoding.t()) :: binary()

Same as get_type_ids/1, but returns binary with u32 values.

get_word_ids(encoding)

@spec get_word_ids(Encoding.t()) :: [non_neg_integer() | nil]

Get word ids from an encoding.

n_tokens(encoding)

@spec n_tokens(encoding :: Encoding.t()) :: non_neg_integer()

Returns the number of tokens in an Encoding.t().

pad(encoding, target_length, opts \\ [])

@spec pad(Encoding.t(), non_neg_integer(), padding_opts()) :: Encoding.t()

Pad the encoding to the given length.

set_sequence_id(encoding, id)

@spec set_sequence_id(Encoding.t(), non_neg_integer()) :: Encoding.t()

Set the given sequence id for the whole range of tokens contained in this Encoding.

token_to_chars(encoding, token)

@spec token_to_chars(Encoding.t(), non_neg_integer()) ::
  {non_neg_integer(), {non_neg_integer(), non_neg_integer()}} | nil

Get the offsets of the token at the given index.

token_to_sequence(encoding, token)

@spec token_to_sequence(Encoding.t(), non_neg_integer()) :: non_neg_integer() | nil

Returns the index of the sequence containing the given token

token_to_word(encoding, token)

@spec token_to_word(Encoding.t(), non_neg_integer()) ::
  {non_neg_integer(), non_neg_integer()} | nil

Get the word that contains the token at the given index.

truncate(encoding, max_length, opts \\ [])

@spec truncate(Encoding.t(), non_neg_integer(), truncation_opts()) :: Encoding.t()

Truncate the encoding to the given length.

word_to_chars(encoding, word, seq_id)

@spec word_to_chars(Encoding.t(), non_neg_integer(), non_neg_integer()) ::
  {non_neg_integer(), non_neg_integer()} | nil

Get the offsets of the word at the given index in the input sequence.

word_to_tokens(encoding, word, seq_id)

@spec word_to_tokens(Encoding.t(), non_neg_integer(), non_neg_integer()) ::
  {non_neg_integer(), non_neg_integer()} | nil

Get the encoded tokens corresponding to the word at the given index in the input sequence, with the form (start_token, end_token + 1)