Tokenizers.Encoding (Tokenizers v0.1.1)

The struct and associated functions for an encoding, the output of a tokenizer.

Use these functions to retrieve the inputs needed for a natural language processing machine learning model.

Types

t()

Specs

t() :: %Tokenizers.Encoding{reference: reference(), resource: binary()}

Functions

get_attention_mask(encoding)

Specs

get_attention_mask(Encoding.t()) :: [integer()]

Get the attention mask from an encoding.

get_ids(encoding)

Specs

get_ids(Encoding.t()) :: [integer()]

Get the ids from an encoding.

get_offsets(encoding)

Specs

get_offsets(Encoding.t()) :: [{integer(), integer()}]

Get offsets from an encoding.

get_special_tokens_mask(encoding)

Specs

get_special_tokens_mask(Encoding.t()) :: [integer()]

Get special tokens mask from an encoding.

get_tokens(encoding)

Specs

get_tokens(Encoding.t()) :: [binary()]

Get the tokens from an encoding.

get_type_ids(encoding)

Specs

get_type_ids(Encoding.t()) :: [integer()]

Get token type ids from an encoding.

n_tokens(encoding)

Specs

n_tokens(encoding :: Encoding.t()) :: non_neg_integer()

Returns the number of tokens in an Encoding.t().

pad(encoding, length, opts \\ [])

Specs

pad(encoding :: Encoding.t(), length :: pos_integer(), opts :: Keyword.t()) ::
  Encoding.t()

Pad the encoding to the given length.

Options

direction - The padding direction. Can be :right or :left. Default: :right.

pad_id - The id corresponding to the padding token. Default: 0.

pad_token - The padding token to use. Default: "[PAD]".

pad_type_id - The type ID corresponding to the padding token. Default: 0.

truncate(encoding, max_len, opts \\ [])

Specs

truncate(encoding :: Encoding.t(), length :: integer(), opts :: Keyword.t()) ::
  Encoding.t()

Truncate the encoding to the given length.

Options

direction - The truncation direction. Can be :right or :left. Default: :right.

stride - The length of previous content to be included in each overflowing piece. Default: 0.