Tokenizers.PreTokenizer (Tokenizers v0.4.0-dev)

The PreTokenizer takes care of splitting the input according to a set of rules. This pre-processing lets you ensure that the underlying Model does not build tokens across multiple “splits”. For example if you don’t want to have whitespaces inside a token, then you can have a PreTokenizer that splits on these whitespaces.

You can easily combine multiple PreTokenizer together using a Sequence (see below). The PreTokenizer is also allowed to modify the string, just like a Normalizer does. This is necessary to allow some complicated algorithms that require to split before normalizing (e.g. the ByteLevel)

byte_level_opts()

@type byte_level_opts() :: [add_prefix_space: boolean(), use_regex: boolean()]

Options for ByteLevel pre-tokenizer. All values are optional.

:add_prefix_space (default true) - Whether to add a space to the first word if there isn’t already one. This lets us treat hello exactly like say hello.

:use_regex (default true) - Set this to False to prevent this pre_tokenizer from using the GPT2 specific regexp for spliting on whitespace.

digits_opts()

@type digits_opts() :: [{:individual_digits, boolean()}]

Options for Digits pre-tokenizer. All values are optional.

:individual_digits (default false) - Whether to split individual digits or not.

metaspace_opts()

@type metaspace_opts() :: [replacement: char(), add_prefix_space: boolean()]

Options for Metaspace pre-tokenizer. All values are optional.

:replacement (default "▁") - The replacement character to use.

:add_prefix_space (default true) - Whether to add a space to the first word if there isn’t already one. This lets us treat hello exactly like say hello.

spit_opts()

@type spit_opts() :: [{:invert, boolean()}]

Options for Split pre-tokenizer. All values are optional.

:invert (default false) - Whether to invert the split or not.

split_delimiter_behaviour()

@type split_delimiter_behaviour() ::
  :removed | :isolated | :merged_with_previous | :merged_with_next | :contiguous

Specifies how delimiter should behave for several pretokenizers.

t()

@type t() :: %Tokenizers.PreTokenizer{resource: reference()}

bert_pre_tokenizer()

@spec bert_pre_tokenizer() :: t()

Creates BertPreTokenizer pre-tokenizer.

Splits for use in Bert models.

byte_level(opts \\ [])

@spec byte_level(opts :: byte_level_opts()) :: t()

Creates ByteLevel PreTokenizer.

Splits on whitespaces while remapping all the bytes to a set of visible characters. This technique as been introduced by OpenAI with GPT-2 and has some more or less nice properties:

Since it maps on bytes, a tokenizer using this only requires 256 characters as initial alphabet (the number of values a byte can have), as opposed to the 130,000+ Unicode characters.

A consequence of the previous point is that it is absolutely unnecessary to have an unknown token using this since we can represent anything with 256 tokens (Youhou!! 🎉🎉)

For non ascii characters, it gets completely unreadable, but it works nonetheless!

byte_level_alphabet()

@spec byte_level_alphabet() :: charlist()

Gets ByteLevel pre-tokenizer's alphabet.

char_delimiter_split(delimiter)

@spec char_delimiter_split(delimiter :: char()) :: t()

Creates CharDelimiterSplit pre-tokenizer.

This pre-tokenizer simply splits on the provided delimiter. Works almost like the .split(delimiter) function, except that it accounts for multiple consecutive spaces

digits(opts \\ [])

@spec digits(opts :: digits_opts()) :: t()

Creates Digits pre-tokenizer.

Splits the numbers from any other characters.

metaspace(opts \\ [])

@spec metaspace(opts :: metaspace_opts()) :: t()

Creates Metaspace pre-tokenizer.

Splits on whitespaces and replaces them with a special char “▁” (U+2581)

pre_tokenize(normalizer, input)

@spec pre_tokenize(pre_tokenizer :: t(), sequence :: String.t()) ::
  {:ok, [{String.t(), {integer(), integer()}}]}

Converts a string into a sequence of pre-tokens.

punctuation(behavor)

@spec punctuation(behavor :: split_delimiter_behaviour()) :: t()

Creates Punctuation pre-tokenizer.

Will isolate all punctuation characters.

sequence(pre_tokenizers)

@spec sequence(pre_tokenizers :: [t()]) :: t()

Creates Sequence pre-tokenizer.

Lets you compose multiple PreTokenizer that will be run in the given order

split(pattern, behavior, opts \\ [])

@spec split(
  pattern :: String.t(),
  behavior :: split_delimiter_behaviour(),
  opts :: spit_opts()
) :: t()

Creates Split pre-tokenizer.

Versatile pre-tokenizer that splits on provided pattern and according to provided behavior. The pattern can be inverted if necessary.

pattern should be either a custom string or regexp.

behavior should be one of:

:removed

:isolated

:merged_with_previous

:merged_with_next

:contiguous

whitespace()

@spec whitespace() :: t()

Creates Whitespace pre-tokenizer.

Splits on word boundaries (using the following regular expression: w+|[^w ]+

whitespace_split()

@spec whitespace_split() :: t()

Creates WhitespaceSplit pre-tokenizer.

Splits on any whitespace character