Tokenizers.Tokenizer (Tokenizers v0.2.0-dev)

The struct and associated functions for a tokenizer.

A Tokenizers.Tokenizer.t() is a container that holds the constituent parts of the tokenization pipeline.

When you call Tokenizers.Tokenizer.encode/3, the input text goes through the following pipeline:

normalization

pre-tokenization

model

post-processing

This returns a Tokenizers.Encoding.t(), which can then give you the token ids for each token in the input text. These token ids are usually used as the input for natural language processing machine learning models.

Types

encode_input()

Specs

encode_input() :: String.t() | {String.t(), String.t()}

An input being a subject to tokenization.

Can be either a single sequence, or a pair of sequences.

t()

Specs

t() :: %Tokenizers.Tokenizer{reference: reference(), resource: binary()}

Functions

decode(tokenizer, ids, opts \\ [])

Specs

decode(Tokenizer.t(), non_neg_integer() | [non_neg_integer()], Keyword.t()) ::
  {:ok, String.t() | [String.t()]} | {:error, term()}

Decode the given list of ids or list of lists of ids back to strings.

Options

:skip_special_tokens - whether the special tokens should be removed from the decoded string. Defaults to true.

encode(tokenizer, input, opts \\ [])

Specs

encode(Tokenizer.t(), encode_input() | [encode_input()], Keyword.t()) ::
  {:ok, Encoding.t() | [Encoding.t()]} | {:error, term()}

Encode the given sequence or batch of sequences to a Tokenizers.Encoding.t().

Options

:add_special_tokens - whether to add special tokens to the encoding. Defaults to true.

from_file(path)

Specs

from_file(String.t()) :: {:ok, Tokenizer.t()} | {:error, term()}

Instantiate a new tokenizer from the file at the given path.

from_pretrained(identifier)

Specs

from_pretrained(String.t()) :: {:ok, Tokenizer.t()} | {:error, term()}

Instantiate a new tokenizer from an existing file on the Hugging Face Hub.

get_model(tokenizer)

Specs

get_model(Tokenizer.t()) :: Tokenizers.Model.t()

Get the Tokenizer's Model.

get_vocab(tokenizer)

Specs

get_vocab(Tokenizer.t()) :: %{required(binary()) => integer()}

Get the tokenizer's vocabulary as a map of token to id.

get_vocab_size(tokenizer)

Specs

get_vocab_size(Tokenizer.t()) :: non_neg_integer()

Get the number of tokens in the vocabulary.

id_to_token(tokenizer, id)

Specs

id_to_token(Tokenizer.t(), integer()) :: String.t()

Convert a given id to its token.

save(tokenizer, path)

Specs

save(Tokenizer.t(), String.t()) :: {:ok, String.t()} | {:error, term()}

Save the tokenizer to the provided path.

token_to_id(tokenizer, token)

Specs

token_to_id(Tokenizer.t(), binary()) :: non_neg_integer()

Convert a given token to its id.