Tokenizers.Tokenizer (Tokenizers v0.4.0-dev)

The struct and associated functions for a tokenizer.

A Tokenizers.t() is a container that holds the constituent parts of the tokenization pipeline.

When you call Tokenizers.Tokenizer.encode/3, the input text goes through the following pipeline:

normalization

pre-tokenization

model

post-processing

This returns a Tokenizers.Encoding.t(), which can then give you the token ids for each token in the input text. These token ids are usually used as the input for natural language processing machine learning models.

encode_input()

@type encode_input() :: String.t() | {String.t(), String.t()}

An input being a subject to tokenization.

Can be either a single sequence, or a pair of sequences.

padding_options()

@type padding_options() :: [
  strategy: :batch_longest | {:fixed, non_neg_integer()},
  direction: :left | :right,
  pad_to_multiple_of: non_neg_integer(),
  pad_id: non_neg_integer(),
  pad_type_id: non_neg_integer(),
  pad_token: String.t()
]

Padding options. All options can be ommited.

:strategy (default: :batch_longest) - the strategy to use when padding.

:direction (default: :right) - the direction to use when padding.

:pad_to_multiple_of (default: 0) - the multiple to pad to.

:pad_id (default: 0) - the id of the token to use for padding.

:pad_type_id (default: 0) - the id of the token type to use for padding.

:pad_token (default: "<pad>") - the token to use for padding.

t()

@type t() :: %Tokenizers.Tokenizer{resource: reference()}

truncation_options()

@type truncation_options() :: [
  max_length: non_neg_integer(),
  stride: non_neg_integer(),
  strategy: :longest_first | :only_first | :only_second,
  direction: :left | :right
]

Truncation options. All options can be ommited.

:max_length (default: 512) - the maximum length to truncate the model's input to.

:stride (default: 0) - the stride to use when overflowing the model's input.

:strategy (default: :longest_first) - the strategy to use when overflowing the model's input. *:direction(default::right`) - the direction to use when overflowing the model's input.

add_special_tokens(tokenizer, tokens)

@spec add_special_tokens(tokenizer :: t(), tokens :: [String.t()]) ::
  non_neg_integer()

Adds special tokens to the vocabulary. These tokens are special. To add regular tokens - use add_tokens/2.

add_tokens(tokenizer, tokens)

@spec add_tokens(tokenizer :: t(), tokens :: [String.t()]) :: non_neg_integer()

Adds tokens to the vocabulary. These tokens are not special. To add special tokens - use add_special_tokens/2.

decode(tokenizer, ids, options \\ [])

@spec decode(
  tokenizer :: t(),
  ids :: [non_neg_integer()],
  options :: [{:skip_special_tokens, boolean()}]
) :: {:ok, String.t()} | {:error, term()}

Decodes the given list of ids back to a string.

Options:

:skip_special_tokens (default: true) - whether to remove special tokens from the decoded string.

decode_batch(tokenizer, sentences, options \\ [])

@spec decode_batch(
  tokenizer :: t(),
  sentences :: [[non_neg_integer()]],
  options :: [{:skip_special_tokens, boolean()}]
) :: {:ok, [String.t()]} | {:error, term()}

Decode the given list of ids or list of lists of ids back to strings.

disable_padding(tokenizer)

@spec disable_padding(tokenizer :: t()) :: t()

Disable padding for the tokenizer.

disable_truncation(tokenizer)

@spec disable_truncation(tokenizer :: t()) :: t()

Disable truncation for the tokenizer.

encode(tokenizer, input, options \\ [])

@spec encode(
  tokenizer :: t(),
  input :: encode_input(),
  options :: [{:add_special_tokens, boolean()}]
) :: {:ok, Tokenizers.Encoding.t()} | {:error, term()}

Encode the given sequence to a Tokenizers.Encoding.t().

Options:

:add_special_tokens (default: true) - whether to add special tokens to the sequence.

encode_batch(tokenizer, input, options \\ [])

@spec encode_batch(
  tokenizer :: t(),
  input :: [encode_input()],
  options :: [{:add_special_tokens, boolean()}]
) :: {:ok, [Tokenizers.Encoding.t()]} | {:error, term()}

Encode the given batch of sequences to a Tokenizers.Encoding.t().

For options check encode/3.

from_buffer(data, options \\ [])

@spec from_buffer(
  data :: String.t(),
  keyword()
) :: {:ok, t()} | {:error, term()}

Instantiate a new tokenizer from the buffer.

from_file(path, options \\ [])

@spec from_file(
  path :: String.t(),
  keyword()
) :: {:ok, t()} | {:error, term()}

Instantiate a new tokenizer from the file at the given path.

from_pretrained(identifier, opts \\ [])

@spec from_pretrained(String.t(), Keyword.t()) :: {:ok, t()} | {:error, term()}

Instantiate a new tokenizer from an existing file on the Hugging Face Hub.

This is going to download a tokenizer file, save it to disk and load that file.

Options

:http_client - A tuple with a module and options. This module should implement the request/1 function, accepting a keyword list with the options for a request. This is inspired by Req.request/1: https://hexdocs.pm/req/Req.html#request/1

The default HTTP client config is: {Tokenizers.HTTPClient, []}. Since it's inspired by Req, it's possible to use that client without any adjustments.

When making request, the options :url and :method are going to be overridden. :headers contains the "user-agent" set by default.

:revision - The revision name that should be used for fetching the tokenizers from Hugging Face.

:use_cache - Tells if it should read from cache when the file already exists. Defaults to true.

:cache_dir - The directory where cache is saved. Files are written to cache even if :use_cache is false. By default it uses :filename.basedir/3 to get a cache dir based in the "tokenizers_elixir" application name.

get_decoder(tokenizer)

@spec get_decoder(t()) :: Tokenizers.Decoder.t() | nil

Get the Tokenizer's Decoder.

get_model(tokenizer)

@spec get_model(t()) :: Tokenizers.Model.t()

Get the Tokenizer's Model.

get_normalizer(tokenizer)

@spec get_normalizer(t()) :: Tokenizers.Normalizer.t() | nil

Get the Tokenizer's Normalizer.

get_post_processor(tokenizer)

@spec get_post_processor(t()) :: Tokenizers.PostProcessor.t() | nil

Get the Tokenizer's PostProcessor.

get_pre_tokenizer(tokenizer)

@spec get_pre_tokenizer(t()) :: Tokenizers.PreTokenizer.t() | nil

Get the Tokenizer's PreTokenizer.

get_vocab(tokenizer, with_additional_tokens \\ true)

@spec get_vocab(tokenizer :: t(), with_additional_tokens :: boolean()) :: %{
  required(String.t()) => integer()
}

Get the tokenizer's vocabulary as a map of token to id.

get_vocab_size(tokenizer, with_additional_tokens \\ true)

@spec get_vocab_size(tokenizer :: t(), with_additional_tokens :: boolean()) ::
  non_neg_integer()

Get the number of tokens in the vocabulary.

id_to_token(tokenizer, id)

@spec id_to_token(t(), integer()) :: String.t() | nil

Convert a given id to its token.

init(model)

@spec init(model :: Tokenizers.Model.t()) :: {:ok, t()} | {:error, any()}

Instantiate a new tokenizer from an existing models.

save(tokenizer, path, options \\ [])

Save the tokenizer to the provided path. Options:

:pretty - Whether to pretty print the JSON file. Defaults to true.

set_decoder(tokenizer, decoder)

@spec set_decoder(t(), Tokenizers.Decoder.t()) :: t()

Set the Tokenizer's Decoder.

set_model(tokenizer, model)

@spec set_model(t(), Tokenizers.Model.t()) :: t()

Set the Tokenizer's Model.

set_normalizer(tokenizer, normalizer)

@spec set_normalizer(t(), Tokenizers.Normalizer.t()) :: t()

Set the Tokenizer's Normalizer.

set_padding(tokenizer, opts)

@spec set_padding(tokenizer :: t(), opts :: padding_options()) :: t()

Set padding for the tokenizer.

set_post_processor(tokenizer, post_processor)

@spec set_post_processor(t(), Tokenizers.PostProcessor.t()) :: t()

Set the Tokenizer's PostProcessor.

set_pre_tokenizer(tokenizer, pre_tokenizer)

@spec set_pre_tokenizer(t(), Tokenizers.PreTokenizer.t()) :: t()

Set the Tokenizer's PreTokenizer.

set_truncation(tokenizer, opts \\ [])

@spec set_truncation(
  tokenizer :: t(),
  opts :: truncation_options()
) :: t()

Set truncation for the tokenizer.

token_to_id(tokenizer, token)

@spec token_to_id(t(), String.t()) :: non_neg_integer() | nil

Convert a given token to its id.

train_from_files(tokenizer, files, trainer \\ nil)

@spec train_from_files(
  tokenizer :: t(),
  files :: [String.t()],
  trainer :: Tokenizers.Trainer.t() | nil
) :: {:ok, t()} | {:error, term()}

Train the tokenizer on the given files.